Adapted all of the wave, multiwave, and scripts to be created in defined folder structure.
I will still need to work on implementing the tempfile structure, but I will leave that for later. I want to focus on implementing all of the functions required for the qt3utils. The next step will be to make this an importable libray. Maybe just do this with github, then I can use the colors module from ethanalysis. Think about this some more.