Added functions to convert wavefiles to arrays of waveforms. Also, added to the visulaization module to automatically plot wavefiles from text files, which is really useful for testing all of these programs.
I also added features that will be necessary for implementation of experiment schemes. These were added to the waveforms module, and include functions to pad waveforms with zeros or ones from the right or the left to get to a certain length, and also a more robust function to make waveforms mod 64. Beyond that, I added a function to downconvert the clock frequency so that I can access lower frequencies than the minumum allowed by the dpg11.